+++ domain/src/main/java/com/onewx2m/domain/repository/BuzzzzingLocationRepository.kt
@@ -9,7 +9,7 @@ interface BuzzzzingLocationRepository {
     suspend fun getBuzzzzingLocation(
         cursorId: Int,
         keyword: String?,
-        categoryId: Long?,
+        categoryId: Int?,
         congestionSort: String,
         cursorCongestionLevel: Int?,
     ): Flow<Outcome<BuzzzzingLocation>>
