+++ remote/src/main/java/com/onewx2m/remote/datasource/RemoteBuzzzzingLocationDataSourceImpl.kt
@@ -18,7 +18,7 @@ class RemoteBuzzzzingLocationDataSourceImpl @Inject constructor(
     override suspend fun getBuzzzzingLocation(
         cursorId: Int,
         keyword: String?,
-        categoryId: Long?,
+        categoryId: Int?,
         congestionSort: String,
         cursorCongestionLevel: Int?,
     ): Flow<Outcome<BuzzzzingLocationEntity>> = flow {
